requirements.txt
django==1.11.14
django-tastypie
django-jsonfield
pillow
django-cors-headers